Fix Eat and Create apple

+++ src/apple.c
@@ -4,7 +4,7 @@
 
 void get_random_position_apple(struct Apple *apple)
 {
-    int pos[] = {0, 36, 72, 108, 144, 180, 216, 252, 288, 324, 360, 396, 432, 468, 504, 540, 576, 612, 648, 684, 720, 756, 792, 828, 864};
+    float pos[] = {0, 36, 72, 108, 144, 180, 216, 252, 288, 324, 360, 396, 432, 468, 504, 540, 576, 612, 648, 684, 720, 756, 792, 828, 864};
     apple->x = pos[GetRandomValue(0, 24)];
     apple->y = pos[GetRandomValue(0, 24)];
 }
@@ -13,7 +13,7 @@ void draw_apple(struct Apple *apple)
 {
     Rectangle rect = {(float)apple->x, (float)apple->y, SIZE, SIZE};
     Color red = {250, 56, 42, 255};
-    DrawRectangleRounded(rect, 0.3, 20, red);
+    DrawRectangleRounded(rect, 0.3f, 20, red);
 }
 
 void snake_eats_apple(struct Apple *apple, struct Snake *snake)
